move type declarations to module file ...might (or might not) improve TS intellisense
